Fix network configuration
This changes the order of the WAN and LAN interface to match the OpenWrt
default. This works with the armvirt target. The MIPS malta target still
has the wrong order. the x86 target only uses one Ethernet device.

Print the qemu command in addition.

Wait till the console is available.