ansible: do not use dnsmasq as DNS resolver
Prevent that dnsmasq tries to run as a DNS resolver on port 53.
DNS settings are out of scope of this playbook so don't interfere with
what already has been configured.

On Ubuntu, dnsmasq isn't able to start because systemd-resolved is
running on port 53 by default as local resolver. Trying to turn it off
usually causes more problems than it solves.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>